rename UserRepository to DrillProgressRepository
