+++ stopcovid/status/status.py
@@ -2,7 +2,7 @@ from typing import List
 
 from .drill_instances import DrillInstanceRepository
 from .initiation import DrillInitiator
-from .users import UserRepository
+from .drill_progress import DrillProgressRepository
 from ..dialog.models.events import UserValidated, NextDrillRequested, DialogEventBatch
 
 
@@ -14,7 +14,7 @@ def handle_dialog_event_batches(batches: List[DialogEventBatch]):
         if initiates_first_drill(batch):
             initiator.trigger_first_drill(batch.phone_number, str(batch.batch_id))
 
-    user_repo = UserRepository()
+    user_repo = DrillProgressRepository()
     drill_instance_repo = DrillInstanceRepository()
     for batch in batches:
         user_id = user_repo.update_user(batch)
